+++ src/js/defaults.js
@@ -1,4 +1,8 @@
-import { DRAG_MODE_CROP } from './constants';
+import {
+  DRAG_MODE_CROP,
+  MIN_CONTAINER_HEIGHT,
+  MIN_CONTAINER_WIDTH,
+} from './constants';
 
 export default {
   // Define the view mode of the cropper
@@ -87,8 +91,8 @@ export default {
   minCanvasHeight: 0,
   minCropBoxWidth: 0,
   minCropBoxHeight: 0,
-  minContainerWidth: 200,
-  minContainerHeight: 100,
+  minContainerWidth: MIN_CONTAINER_WIDTH,
+  minContainerHeight: MIN_CONTAINER_HEIGHT,
 
   // Shortcuts of events
   ready: null,